backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for Cloud AI Bank Onboarding"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.104.1,<1.0.0",
    "uvicorn>=0.24.0,<1.0.0",
    "langchain>=1.0.0",
    "langchain-core>=1.0.0",
    "langchain-classic>=0.1.0",
    "langchain-community>=0.2.0",
    "langchain-ollama>=0.1.0",
    "ollama>=0.3.0",
    "faiss-cpu>=1.7.4,<2.0.0",
    "sentence-transformers>=2.2.2,<3.0.0",
    "pymupdf>=1.23.0,<2.0.0",
    "streamlit>=1.29.0,<2.0.0",
    "requests>=2.31.0,<3.0.0",
    "pydantic>=2.5.0,<3.0.0"
]

[dependency-groups]
dev = [
    "pytest>=7.4.3,<8.0.0",
    "black>=23.12.0,<24.0.0",
    "ruff>=0.1.7,<1.0.0",
    "ipykernel>=6.23.0,<7.0.0",
    "jupyter>=1.0.0,<2.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]
